requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
astor==0.8.1
attrdict==2.0.1
attrs==25.4.0
babel==2.18.0
bce-python-sdk==0.9.60
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cryptography==46.0.5
cssselect==1.4.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.4
decorator==5.2.1
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fire==0.7.1
Flask==3.1.3
flask-babel==4.0.0
fonttools==4.61.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.68.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
groq==1.1.1
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
lmdb==1.7.5
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
more-itertools==10.8.0
narwhals==2.18.0
networkx==3.4.2
numpy==2.2.6
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==26.0
paddleocr==2.7.3
paddlepaddle==3.0.0
pandas==2.3.3
pandas-stubs==2.3.3.260113
pdf2docx==0.5.10
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
premailer==3.10.0
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyMuPDF==1.27.1
pyparsing==3.3.2
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
RapidFuzz==3.14.3
rarfile==4.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.2
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
streamlit==1.55.0
tenacity==9.1.4
termcolor==3.3.0
tifffile==2025.5.10
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
types-pytz==2025.2.0.20251108
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
visualdl==2.5.3
websockets==16.0
Werkzeug==3.1.6
